Move pagination block logic to Trait
